Changed size of component into ComponentSize class. Implemented filtration with any type of filter for GPU via CriteriaQuery. Added partial update of GPU via reflection. Added mappers (GPUMapper and GPUFilterMapper)

+++ src/main/java/com/pcplanet/pcplanetbackend/component/gpu/output_interface/GPUOutputInterface.java
@@ -11,9 +11,8 @@ import lombok.NoArgsConstructor;
 public class GPUOutputInterface {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Short id;
+    private Integer id;
     private String interfaceName;
-
     public GPUOutputInterface(String interfaceName) {
         this.interfaceName = interfaceName;
     }
